TD reorganize tests
